+++ src/components/Navbar/Navbar.jsx
@@ -1,12 +1,15 @@
 import React from 'react';
 
+import LogoutIcon from '@mui/icons-material/Logout';
 import MenuIcon from '@mui/icons-material/Menu';
 import PersonIcon from '@mui/icons-material/Person';
 import { Box, IconButton, Toolbar, Typography } from '@mui/material';
 
 import { ThemeToggler } from '../';
+import { useAuth } from '../../context';
 
 const Navbar = ({ isSidebarClose, setSidebarClose }) => {
+  const { signout } = useAuth();
   return (
     <>
       <Toolbar
@@ -17,14 +20,12 @@ const Navbar = ({ isSidebarClose, setSidebarClose }) => {
           display: 'flex',
           alignItems: 'center',
           justifyContent: 'space-between',
-        }}
-      >
+        }}>
         <Box sx={{ display: 'flex', alignItems: 'center' }}>
           {isSidebarClose && (
             <IconButton
               onClick={() => setSidebarClose(false)}
-              sx={{ display: { xs: 'none', sm: 'block' } }}
-            >
+              sx={{ display: { xs: 'none', sm: 'block' } }}>
               <MenuIcon />
             </IconButton>
           )}
@@ -36,6 +37,9 @@ const Navbar = ({ isSidebarClose, setSidebarClose }) => {
           <ThemeToggler />
           <IconButton>
             <PersonIcon fontSize='large' />
+          </IconButton>
+          <IconButton onClick={signout}>
+            <LogoutIcon fontSize='large' />
           </IconButton>
         </Box>
       </Toolbar>
