+++ src/context/index.js
@@ -1,1 +1,2 @@
 export { DarkModeProvider, useDarkMode } from './DarkMode';
+export { AuthProvider, useAuth } from './Auth';
