+++ src/main.jsx
@@ -6,8 +6,7 @@ import { ThemeProvider } from '@mui/material/styles';
 import { registerSW } from 'virtual:pwa-register';
 
 import Routes from './Routes';
-import { DarkModeProvider } from './context';
-import { AuthProvider } from './context/AuthContext';
+import { AuthProvider, DarkModeProvider } from './context';
 import './index.css';
 import { theme } from './styles/theme';
 
@@ -15,13 +14,13 @@ registerSW();
 
 ReactDOM.render(
   <React.StrictMode>
-    <AuthProvider>
-      <ThemeProvider theme={theme}>
-        <DarkModeProvider>
+    <ThemeProvider theme={theme}>
+      <DarkModeProvider>
+        <AuthProvider>
           <Routes />
-        </DarkModeProvider>
-      </ThemeProvider>
-    </AuthProvider>
+        </AuthProvider>
+      </DarkModeProvider>
+    </ThemeProvider>
   </React.StrictMode>,
   document.getElementById('root'),
 );
