+++ src/pages/AuthPage/SignUp.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
 import { Avatar, Box, Button, Container, Grid, Paper, TextField, Typography } from '@mui/material';
 
-import { useAuth } from '../../context/AuthContext';
+import { useAuth } from '../../context';
 import GridImage from './components/GridImage';
 
 const SignUp = () => {
